+++ src/main/java/pl/allegro/finance/tradukisto/MoneyConverters.java
@@ -1,12 +1,21 @@
 package pl.allegro.finance.tradukisto;
 
+import pl.allegro.finance.tradukisto.internal.BigDecimalToStringConverter;
+
+import java.math.BigDecimal;
+
 import static com.google.common.base.Verify.verifyNotNull;
 import static pl.allegro.finance.tradukisto.internal.Container.brazilianPortugueseContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.czechContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.englishContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.frenchContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.germanContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.italianContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.kazakhContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.latvianContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.polishContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.russianContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.serbianContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.slovakContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.latvianContainer;
 import static pl.allegro.finance.tradukisto.internal.Container.kazakhContainer;
@@ -24,12 +33,15 @@ public enum MoneyConverters {
     GERMAN_BANKING_MONEY_VALUE(germanContainer().getBankingMoneyConverter()),
     RUSSIAN_BANKING_MONEY_VALUE(russianContainer().getBankingMoneyConverter()),
     POLISH_BANKING_MONEY_VALUE(polishContainer().getBankingMoneyConverter()),
+    ITALIAN_BANKING_MONEY_VALUE(italianContainer().getBankingMoneyConverter()),
     CZECH_BANKING_MONEY_VALUE(czechContainer().getBankingMoneyConverter()),
     SLOVAK_BANKING_MONEY_VALUE(slovakContainer().getBankingMoneyConverter()),
     LATVIAN_BANKING_MONEY_VALUE(latvianContainer().getBankingMoneyConverter()),
     KAZAKH_BANKING_MONEY_VALUE(kazakhContainer().getBankingMoneyConverter()),
     UKRAINIAN_BANKING_MONEY_VALUE(ukrainianContainer().getBankingMoneyConverter()),
-    TURKISH_BANKING_MONEY_VALUE(turkishContainer().getBankingMoneyConverter());
+    TURKISH_BANKING_MONEY_VALUE(turkishContainer().getBankingMoneyConverter()),
+    SERBIAN_BANKING_MONEY_VALUE(serbianContainer().getBankingMoneyConverter()),
+    FRENCH_BANKING_MONEY_VALUE(frenchContainer().getBankingMoneyConverter());
 
     private final BigDecimalToStringConverter converter;