+++ src/main/java/pl/allegro/finance/tradukisto/internal/Container.java
@@ -7,9 +7,14 @@ import pl.allegro.finance.tradukisto.internal.languages.czech.CzechIntegerToWord
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechValues;
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechValuesForSmallNumbers;
 import pl.allegro.finance.tradukisto.internal.languages.english.EnglishValues;
+import pl.allegro.finance.tradukisto.internal.languages.french.FrenchIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.french.FrenchValues;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanIntegerToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanThousandToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanValues;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianThousandToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianValues;
 import pl.allegro.finance.tradukisto.internal.languages.kazakh.KazakhValues;
 import pl.allegro.finance.tradukisto.internal.languages.latvian.LatvianValues;
 import pl.allegro.finance.tradukisto.internal.languages.polish.PolishValues;
@@ -18,6 +23,7 @@ import pl.allegro.finance.tradukisto.internal.languages.portuguese.PortugueseInt
 import pl.allegro.finance.tradukisto.internal.languages.portuguese.PortugueseIntegerToWordsConverterAdapter;
 import pl.allegro.finance.tradukisto.internal.languages.portuguese.PortugueseThousandToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.russian.RussianValues;
+import pl.allegro.finance.tradukisto.internal.languages.serbian.SerbianValues;
 import pl.allegro.finance.tradukisto.internal.languages.slovak.SlovakValues;
 import pl.allegro.finance.tradukisto.internal.languages.slovak.SlovakValuesForSmallNumbers;
 import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishBigDecimalToBankingMoneyConverter;
@@ -34,6 +40,10 @@ public final class Container {
 
     public static Container russianContainer() {
         return new Container(new RussianValues());
+    }
+
+    public static Container serbianContainer() {
+        return new Container(new SerbianValues());
     }
 
     public static Container ukrainianContainer() {
@@ -73,6 +83,23 @@ public final class Container {
         return new Container(new EnglishValues());
     }
 
+    public static Container frenchContainer() {
+
+        FrenchValues values = new FrenchValues();
+
+        HundredsToWordsConverter hundredsToWordsConverter =
+                new HundredsToWordsConverter(values.baseNumbers(), values.twoDigitsNumberSeparator());
+        IntegerToWordsConverter frenchIntegerToWordsConverter =
+                new IntegerToWordsConverter(hundredsToWordsConverter, values.pluralForms());
+        IntegerToStringConverter converter =
+                new FrenchIntegerToWordsConverter(frenchIntegerToWordsConverter, values.exceptions(), values.pluralForms());
+
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter =
+                new BigDecimalToBankingMoneyConverter(converter, values.currency());
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
     public static Container germanContainer() {
 
         GermanValues values = new GermanValues();
@@ -83,6 +110,22 @@ public final class Container {
         IntegerToStringConverter converter = new GermanIntegerToWordsConverter(
                 new IntegerToWordsConverter(germanThousandToWordsConverter, values.pluralForms()), values.exceptions(),
                 germanThousandToWordsConverter);
+
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter = new BigDecimalToBankingMoneyConverter(
+                converter, values.currency());
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
+    public static Container italianContainer() {
+        ItalianValues values = new ItalianValues();
+
+        ItalianThousandToWordsConverter italianThousandToWordsConverter = new ItalianThousandToWordsConverter(
+                values.baseNumbers());
+
+        IntegerToStringConverter converter = new ItalianIntegerToWordsConverter(
+                new IntegerToWordsConverter(italianThousandToWordsConverter, values.pluralForms()), values.exceptions(),
+                italianThousandToWordsConverter);
 
         BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter = new BigDecimalToBankingMoneyConverter(
                 converter, values.currency());