+++ src/test/groovy/pl/allegro/finance/tradukisto/MoneyConvertersTest.groovy
@@ -6,6 +6,7 @@ import spock.lang.Specification
 import static pl.allegro.finance.tradukisto.MoneyConverters.CZECH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.ENGLISH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.GERMAN_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.ITALIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.KAZAKH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.LATVIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.POLISH_BANKING_MONEY_VALUE
@@ -13,6 +14,8 @@ import static pl.allegro.finance.tradukisto.MoneyConverters.BRAZILIAN_PORTUGUESE
 import static pl.allegro.finance.tradukisto.MoneyConverters.RUSSIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.SLOVAK_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.UKRAINIAN_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.SERBIAN_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.FRENCH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.TURKISH_BANKING_MONEY_VALUE
 
 class MoneyConvertersTest extends Specification {
@@ -25,6 +28,11 @@ class MoneyConvertersTest extends Specification {
     def "should convert money in German"() {
         expect:
         GERMAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "eintausendzweihundertvierunddreißig € 56/100"
+    }
+
+    def "should convert money in Italian"() {
+        expect:
+        ITALIAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "milleduecentotrentaquattro € 56/100"
     }
 
     def "should convert money in Russian"() {
@@ -65,6 +73,16 @@ class MoneyConvertersTest extends Specification {
     def "should convert money in Ukrainian"() {
         expect:
         UKRAINIAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "одна тисяча двісті тридцять чотири ₴ 56/100"
+    }
+
+    def "should convert money in Serbian"() {
+        expect:
+        SERBIAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "jedna hiljada dvesta trideset četiri RSD 56/100"
+    }
+
+    def "should convert numbers in French"() {
+        expect:
+        FRENCH_BANKING_MONEY_VALUE.asWords(1_234.56) == "mille deux cent trente-quatre € 56/100"
     }
 
     def "should throw exception when null given"() {